requirements.txt
astroid==2.4.2
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.19.4
opencv-contrib-python==4.4.0.46
Pillow==8.0.1
pylint==2.6.0
six==1.15.0
toml==0.10.2
wrapt==1.12.1
